gonna fix blocks array / return type tommorow

+++ io.h
@@ -10,6 +10,6 @@
 #define FILE_ERROR    (-3)
 #define INPUT_ERROR   (-4)
 
-int ***get_block_data(char *file_name);
+unsigned short int ***get_block_data(char *file_name);
 
 #endif //IO_H
